requirements.txt

selenium
webdriver-manager
beautifulsoup4
lxml
pandas
boto3
python-dotenv
psycopg2-binary
duckdb
